requirements.txt
streamlit
pandas
numpy
scikit-learn==1.6.1
google-generativeai
supabase
pyarabic
nltk